+++ backend/apid/graphql/event.go
@@ -24,6 +24,18 @@ func (r *eventImpl) ID(p graphql.ResolveParams) (string, error) {
 	return globalid.EventTranslator.EncodeToString(p.Source), nil
 }
 
+// Namespace implements response to request for 'namespace' field.
+func (r *eventImpl) Namespace(p graphql.ResolveParams) (string, error) {
+	event := p.Source.(*types.Event)
+	if event.Entity != nil {
+		return event.Entity.Namespace, nil
+	}
+	if event.Check != nil {
+		return event.Check.Namespace, nil
+	}
+	return "", nil
+}
+
 // Timestamp implements response to request for 'timestamp' field.
 func (r *eventImpl) Timestamp(p graphql.ResolveParams) (time.Time, error) {
 	event := p.Source.(*types.Event)