+++ dashboard/src/apollo/resolvers/lastNamespace.js
@@ -1,0 +1,18 @@
+export default {
+  defaults: {
+    lastNamespace: "default",
+  },
+  resolvers: {
+    Mutation: {
+      setLastNamespace: (_, { name }, { cache }) => {
+        cache.writeData({
+          data: {
+            lastNamespace: name,
+          },
+        });
+
+        return null;
+      },
+    },
+  },
+};