+++ dashboard/src/apollo/stateLink.js
@@ -1,7 +1,7 @@
 import merge from "deepmerge";
 import { withClientState } from "apollo-link-state";
 import auth from "./resolvers/auth";
-import lastEnvironment from "./resolvers/lastEnvironment";
+import lastNamespace from "./resolvers/lastNamespace";
 import addDeletedFieldTo from "./resolvers/deleted";
 
 const resolvers = merge.all([
@@ -12,7 +12,7 @@ const resolvers = merge.all([
   addDeletedFieldTo("Event"),
   addDeletedFieldTo("Silenced"),
   addDeletedFieldTo("CheckConfig"),
-  lastEnvironment,
+  lastNamespace,
 ]);
 
 const stateLink = ({ cache }) => withClientState({ ...resolvers, cache });